move diagnostics struct & code to its own file
added ToStringMap() method too, for use in main_test.go

should probably do the same with config at some point but not this PR :)
