+++ cmd/root.go
@@ -43,24 +43,11 @@ type Config struct {
 	Verbose bool   `json:"verbose"`
 }
 
-// Diagnostics is a place to put things that are useful for testing and
-// diagnosing issues with otel-cli. The only user-facing feature that should be
-// using these is otel-cli status.
-type Diagnostics struct {
-	IsRecording       bool   `json:"is_recording"`
-	ConfigFileLoaded  bool   `json:"config_file_loaded"`
-	NumArgs           int    `json:"number_of_args"`
-	DetectedLocalhost bool   `json:"detected_localhost"`
-	ParsedTimeoutMs   int64  `json:"parsed_timeout_ms"`
-	OtelError         string `json:"otel_error"`
-}
-
 const defaultOtlpEndpoint = "localhost:4317"
 const spanBgSockfilename = "otel-cli-background.sock"
 
 var exitCode int
 var config Config
-var diagnostics Diagnostics
 
 // rootCmd represents the base command when called without any subcommands
 var rootCmd = &cobra.Command{
